pyproject.toml
[tool.poetry]
name = "pyexcel-xlsxwx"
version = "0.1.5"
description = "Save pyexcel data with XlsxWriter, while retaining good formatting."
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
license = "MIT"
readme = 'README.md'

repository = "https://github.com/patarapolw/pyexcel-xlsxwx"
homepage = "https://github.com/patarapolw/pyexcel-xlsxwx"

keywords = ['pyexcel', 'excel', 'xlsxwriter']

[tool.poetry.dependencies]
python = ">=3.5"
xlsxwriter = "^1.0"
"ruamel.yaml" = "^0.15.50"
importlib_resources = "^1.0"

[tool.poetry.dev-dependencies]
pyexcel = "^0.5.8"
pyexcel-xls = "^0.5.7"
pytest = "^3.7"
